third_party/blink/web_tests/external/wpt/html/dom/elements/global-attributes/custom-attrs-expected.txt

This is a testharness.js-based test.
FAIL Setting an Element's dataset property should not interfere with namespaced attributes with same name assert_equals: expected 3 but got 2
Harness: the test ran to completion.
